test(web): delete dead api test helpers
