--- turbo/apps/web/src/__tests__/api-test-helpers/credits.ts
@@ -1,43 +1,0 @@
-// ---------------------------------------------------------------------------
-// Re-exports: DB-direct seeders.
-//
-// These functions live in db-test-seeders/credits.ts but are re-exported
-// here for backward compatibility — existing test files import from
-// api-test-helpers and should continue to work unchanged.
-// ---------------------------------------------------------------------------
-
-export {
-  updateOrgStripeFields,
-  updateOrgAutoRecharge,
-  updateOrgStripeSubscription,
-  insertCreditExpiresRecord,
-  insertOrgPromoRedemption,
-  insertTestModelUsageEvent,
-  insertTestModelUsageEventForRun,
-  insertTestUsagePricing,
-  deleteTestUsagePricing,
-  seedRealtimeBillingPricing,
-  insertTestUsageEvent,
-  deleteTestUsageEventsByProvider,
-  setTestUsageEventCreatedAt,
-  seedUsageEventRecord,
-  seedInsightsDaily,
-  createCompletedRun,
-} from "../db-test-seeders/credits";
-
-// ---------------------------------------------------------------------------
-// Re-exports: Assertion helpers.
-// ---------------------------------------------------------------------------
-
-export {
-  getOrgBillingFields,
-  getOrgAutoRechargeFields,
-  findCreditExpiresRecords,
-  findCreditExpiresRecordByStripeInvoiceId,
-  findOrgPromoRedemption,
-  findTestUsageEvent,
-  findTestUsageEventsByRunId,
-  findTestRunlessUsageEventsByOrgProvider,
-  findUsageDaily,
-  findInsightsDaily,
-} from "../db-test-assertions/credits";
