--- turbo/apps/web/src/__tests__/api-test-helpers/export.ts
@@ -1,15 +1,0 @@
-// ---------------------------------------------------------------------------
-// Re-exports: DB-direct seeders.
-//
-// These functions live in db-test-seeders/export.ts but are re-exported
-// here for backward compatibility — existing test files import from
-// api-test-helpers and should continue to work unchanged.
-// ---------------------------------------------------------------------------
-
-export { insertTestExportJob } from "../db-test-seeders/export";
-
-// ---------------------------------------------------------------------------
-// Re-exports: Assertion helpers.
-// ---------------------------------------------------------------------------
-
-export { findTestExportJobById } from "../db-test-assertions/export";
