--- turbo/apps/web/src/__tests__/api-test-helpers/index.ts
@@ -1,183 +1,0 @@
-// Barrel re-export for api-test-helpers sub-modules.
-// All existing imports from api-test-helpers continue to work unchanged.
-export {
-  createTestRequest,
-  createDefaultComposeConfig,
-  getTestAuthContext,
-  type ComposeConfigOptions,
-} from "./core";
-export {
-  createTestOrg,
-  deleteOrgRow,
-  updateOrgTier,
-  updateOrgDefaultAgent,
-  setDefaultAgentByComposeId,
-  deleteOrgCacheEntry,
-  insertOrgMembersEntry,
-  insertOrgDefaultModelProvider,
-  insertOrgNonDefaultModelProvider,
-  insertOrgMultiAuthModelProvider,
-  insertUserDefaultModelProvider,
-  insertUserMultiAuthModelProvider,
-  insertUserNonDefaultModelProvider,
-  insertOrgModelPolicy,
-  insertUserModelPreference,
-  deleteTestModelProvider,
-  setOrgCredits,
-  lockOrgAndSetCredits,
-  getOrgDefaultAgent,
-  getOrgCacheEntry,
-  getOrgMembersEntry,
-  countOrgRows,
-  getOrgCredits,
-  insertOrgCacheEntry,
-  ensureOrgRow,
-  insertOrgMembersCacheEntry,
-  findOrgMembersCacheEntry,
-  clearOrgMembersCacheEntry,
-  findTestModelProviderTokenState,
-  findTestOrgModelProviderByType,
-  setTestModelProviderTokenExpiresAt,
-  setTestModelProviderNeedsReconnect,
-  ORG_SENTINEL_USER_ID,
-} from "./org";
-export {
-  createTestComposeVersion,
-  ensureZeroAgentRow,
-  setTestZeroAgentModelProvider,
-  createTestAgentSession,
-  createTestSessionWithConversation,
-  insertTestChatThread,
-  deleteTestChatThread,
-  insertTestChatMessage,
-  getTestChatMessagesByThread,
-  addTestRunToThread,
-  insertTestAssistantEventMessages,
-  setTestChatMessageAttachFiles,
-  setTestChatMessageContent,
-  getTestAgentSessionArtifacts,
-  setTestChatThreadLastReadAt,
-  setTestChatThreadLastReadMessageId,
-  setTestChatThreadModelPin,
-  setTestChatThreadPinnedAt,
-  setTestChatThreadRenamedAt,
-  setTestChatThreadDraft,
-  getTestAgentSessionWithConversation,
-  getTestAgentComposeName,
-  getTestChatThreadLastReadAt,
-  getTestChatThreadLastReadMessageId,
-  getTestChatThreadPinnedAt,
-  getTestChatThreadRenamedAt,
-  getTestUserMessageRunStorage,
-  createTestCompose,
-  updateTestChatThreadTitle,
-} from "./agents";
-export {
-  markRunningRunsAsCompleted,
-  setTestRunStatus,
-  setTestRunResult,
-  setTestRunModelProvider,
-  setTestRunSelectedModel,
-  setTestRunModelProviderMetadata,
-  insertTestZeroRun,
-  expireQueueEntry,
-  insertTestQueueEntry,
-  createTestCallback,
-  linkRunToSchedule,
-  insertTestConversation,
-  insertTestSandboxTelemetry,
-  insertTestUsageDaily,
-  findTestRunsByUserAndPrompt,
-  findTestRunsByUserAndPromptContaining,
-  findTestRunRecord,
-  findTestCheckpoint,
-  findTestZeroRun,
-  findTestRunCallbacks,
-  findTestQueueEntry,
-  findTestCallbacksByRunId,
-  findMostRecentRunForUser,
-  findTestSandboxTelemetry,
-  createTestRun,
-  getTestRun,
-  createTestCheckpoint,
-  completeTestRun,
-  failTestRun,
-} from "./runs";
-export { seedTestSchedule } from "./schedules";
-export {
-  createTestVolumeForOrg,
-  insertStorageVersion,
-  insertTestArtifactStorage,
-  insertTestStorage,
-  insertTestStorageVersion,
-  findTestStorageByName,
-  findTestStorage,
-  findTestSystemStorageByName,
-  getStorageVersionLineage,
-  createTestArtifact,
-  createTestVolume,
-} from "./storage";
-export { createTestSecret, createTestVariable } from "./secrets";
-export {
-  createTestUserConnector,
-  insertTestConnectorSecret,
-  createTestConnectorSession,
-  findTestConnectorSecret,
-  findTestConnectorTokenExpiresAt,
-  createTestConnector,
-} from "./connectors";
-export { createSignedCallbackRequest } from "./callbacks";
-export {
-  updateOrgStripeFields,
-  updateOrgAutoRecharge,
-  updateOrgStripeSubscription,
-  insertCreditExpiresRecord,
-  insertOrgPromoRedemption,
-  insertTestModelUsageEvent,
-  insertTestModelUsageEventForRun,
-  insertTestUsagePricing,
-  deleteTestUsagePricing,
-  seedRealtimeBillingPricing,
-  insertTestUsageEvent,
-  deleteTestUsageEventsByProvider,
-  setTestUsageEventCreatedAt,
-  seedUsageEventRecord,
-  seedInsightsDaily,
-  createCompletedRun,
-  getOrgBillingFields,
-  getOrgAutoRechargeFields,
-  findCreditExpiresRecords,
-  findCreditExpiresRecordByStripeInvoiceId,
-  findOrgPromoRedemption,
-  findTestUsageEvent,
-  findTestUsageEventsByRunId,
-  findTestRunlessUsageEventsByOrgProvider,
-  findUsageDaily,
-  findInsightsDaily,
-} from "./credits";
-export { findTestConnectorBillingByRunId } from "./connector-billing";
-export {
-  insertTestUser,
-  seedUserCacheEntry,
-  insertUserCacheEntry,
-  countUserRows,
-  getPushSubscriptionsByEndpoint,
-  createTestPushSubscription,
-} from "./users";
-export {
-  seedTestSkill,
-  reseedSkills,
-  setAllTestSkillsCommitSha,
-  bindCustomSkillToAgent,
-  createTestZeroSkill,
-  findTestSkillByUrl,
-  getAgentCustomSkills,
-} from "./skills";
-export {
-  createTestRunnerJob,
-  insertTestRunnerState,
-  deleteAllTestRunnerState,
-  findTestRunnerJobEntry,
-} from "./runner";
-export { insertTestExportJob, findTestExportJobById } from "./export";
-export { seedUserFeatureSwitches } from "../db-test-seeders/feature-switches";
