--- turbo/apps/web/src/__tests__/api-test-helpers/org.ts
@@ -1,47 +1,0 @@
-// Re-exports: DB-direct seeders
-export {
-  createTestOrg,
-  deleteOrgRow,
-  updateOrgTier,
-  updateOrgDefaultAgent,
-  setDefaultAgentByComposeId,
-  deleteOrgCacheEntry,
-  insertOrgMembersEntry,
-  insertOrgDefaultModelProvider,
-  insertOrgNonDefaultModelProvider,
-  insertOrgMultiAuthModelProvider,
-  insertUserDefaultModelProvider,
-  insertUserMultiAuthModelProvider,
-  insertUserNonDefaultModelProvider,
-  insertOrgModelPolicy,
-  insertUserModelPreference,
-  deleteTestModelProvider,
-  setOrgCredits,
-  lockOrgAndSetCredits,
-} from "../db-test-seeders/org";
-
-// Re-exports: read-only assertions
-export {
-  getOrgDefaultAgent,
-  getOrgCacheEntry,
-  getOrgMembersEntry,
-  countOrgRows,
-  getOrgCredits,
-  findTestModelProviderTokenState,
-  findTestOrgModelProviderByType,
-  setTestModelProviderTokenExpiresAt,
-  setTestModelProviderNeedsReconnect,
-} from "../db-test-assertions/org";
-
-// Re-exports: test-helpers infrastructure
-export { insertOrgCacheEntry, ensureOrgRow } from "../test-helpers";
-
-// Re-exports: org-members-cache seeders (already migrated)
-export {
-  insertOrgMembersCacheEntry,
-  findOrgMembersCacheEntry,
-  clearOrgMembersCacheEntry,
-} from "../db-test-seeders/org-members-cache";
-
-// Re-exports: constants
-export { ORG_SENTINEL_USER_ID } from "../test-constants/org";
