--- turbo/apps/web/src/__tests__/api-test-helpers/runner.ts
@@ -1,19 +1,0 @@
-// ---------------------------------------------------------------------------
-// Re-exports: DB-direct seeders.
-//
-// These functions live in db-test-seeders/runner.ts but are re-exported
-// here for backward compatibility — existing test files import from
-// api-test-helpers and should continue to work unchanged.
-// ---------------------------------------------------------------------------
-
-export {
-  createTestRunnerJob,
-  insertTestRunnerState,
-  deleteAllTestRunnerState,
-} from "../db-test-seeders/runner";
-
-// ---------------------------------------------------------------------------
-// Re-exports: Assertion helpers.
-// ---------------------------------------------------------------------------
-
-export { findTestRunnerJobEntry } from "../db-test-assertions/runner";
