--- turbo/apps/web/src/__tests__/api-test-helpers/schedules.ts
@@ -1,1 +1,0 @@
-export { seedTestSchedule } from "../db-test-seeders/schedules";
