--- turbo/apps/web/src/__tests__/api-test-helpers/skills.ts
@@ -1,20 +1,0 @@
-// ---------------------------------------------------------------------------
-// Re-exports: DB-direct seeders and assertion helpers.
-//
-// These functions were moved to dedicated directories but are re-exported
-// here for backward compatibility — existing test files import from
-// api-test-helpers and should continue to work unchanged.
-// ---------------------------------------------------------------------------
-
-export {
-  seedTestSkill,
-  reseedSkills,
-  setAllTestSkillsCommitSha,
-  bindCustomSkillToAgent,
-  createTestZeroSkill,
-} from "../db-test-seeders/skills";
-
-export {
-  findTestSkillByUrl,
-  getAgentCustomSkills,
-} from "../db-test-assertions/skills";
